optimize the definition and function of quota

+++ amoro-ams/src/main/java/org/apache/amoro/server/table/DefaultOptimizingState.java
@@ -517,7 +517,7 @@ public class DefaultOptimizingState extends StatedPersistentBase implements Proc
     return tableConfiguration.getOptimizingConfig().isEnabled();
   }
 
-  public Double getTargetQuota() {
+  public int getTargetQuota() {
     return tableConfiguration.getOptimizingConfig().getTargetQuota();
   }
